added a Add account menu on top of the account list
