+++ packages/kit/src/views/AccountManagerStacks/components/WalletEdit/WalletEditButton.tsx
@@ -5,15 +5,11 @@ import { useIntl } from 'react-intl';
 import { ActionList, Divider } from '@onekeyhq/components';
 import { AccountSelectorProviderMirror } from '@onekeyhq/kit/src/components/AccountSelector';
 import { ListItem } from '@onekeyhq/kit/src/components/ListItem';
-import {
-  useAccountSelectorContextData,
-  useActiveAccount,
-} from '@onekeyhq/kit/src/states/jotai/contexts/accountSelector';
+import { useAccountSelectorContextData } from '@onekeyhq/kit/src/states/jotai/contexts/accountSelector';
 import type { IDBWallet } from '@onekeyhq/kit-bg/src/dbs/local/types';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import accountUtils from '@onekeyhq/shared/src/utils/accountUtils';
 
-import { BatchCreateAccountButton } from './BatchCreateAccountButton';
 import { DeviceManagementButton } from './DeviceManagementButton';
 import { HdWalletBackupButton } from './HdWalletBackupButton';
 import { WalletBoundReferralCodeButton } from './WalletBoundReferralCodeButton';
@@ -28,7 +24,6 @@ function WalletEditButtonView({
 }) {
   const intl = useIntl();
   const { config } = useAccountSelectorContextData();
-  const { activeAccount } = useActiveAccount({ num: num ?? 0 });
 
   const showDeviceManagementButton = useMemo(() => {
     return (
@@ -66,23 +61,6 @@ function WalletEditButtonView({
     return accountUtils.isHdWallet({ walletId: wallet?.id });
   }, [wallet]);
 
-  // qr wallet can not batch create account
-  const canBatchCreateAccount = useMemo(() => {
-    if (accountUtils.isQrWallet({ walletId: wallet?.id })) {
-      return false;
-    }
-    if (
-      accountUtils.isHwOrQrWallet({ walletId: wallet?.id }) &&
-      wallet?.isMocked
-    ) {
-      return false;
-    }
-    return (
-      accountUtils.isHdWallet({ walletId: wallet?.id }) ||
-      accountUtils.isHwOrQrWallet({ walletId: wallet?.id })
-    );
-  }, [wallet]);
-
   const estimatedContentHeight = useCallback(async () => {
     let basicHeight = 12;
     if (showDeviceManagementButton) {
@@ -100,32 +78,6 @@ function WalletEditButtonView({
     if (showBackupButton) {
       basicHeight += 44;
     }
-    if (canBatchCreateAccount) {
-      basicHeight += 44;
-    }
-    // const exportKeysVisible = await getExportKeysVisible();
-    // if (exportKeysVisible?.showExportPrivateKey) {
-    //   basicHeight += 44;
-    // }
-
-    // if (exportKeysVisible?.showExportPublicKey) {
-    //   basicHeight += 44;
-    // }
-
-    // if (exportKeysVisible?.showExportMnemonic) {
-    //   basicHeight += 44;
-    // }
-
-    // if (
-    //   firstIndexedAccount?.id !== indexedAccount?.id ||
-    //   firstAccount?.id !== account?.id
-    // ) {
-    //   basicHeight += 44;
-    // }
-
-    // if (showRemoveButton) {
-    //   basicHeight += 54;
-    // }
     return basicHeight;
   }, [
     showDeviceManagementButton,
@@ -133,7 +85,6 @@ function WalletEditButtonView({
     showRemoveWalletButton,
     showBoundReferralCodeButton,
     showBackupButton,
-    canBatchCreateAccount,
   ]);
 
   const renderItems = useCallback(
@@ -163,20 +114,9 @@ function WalletEditButtonView({
           ) : null}
 
           {showDeviceManagementButton ? (
-            <DeviceManagementButton
-              wallet={wallet}
-              onClose={handleActionListClose}
-            />
-          ) : null}
-
-          {canBatchCreateAccount ? (
             <>
-              <BatchCreateAccountButton
-                focusedWalletInfo={{
-                  wallet: wallet as IDBWallet,
-                  device: undefined,
-                }}
-                activeAccount={activeAccount}
+              <DeviceManagementButton
+                wallet={wallet}
                 onClose={handleActionListClose}
               />
               <Divider mx="$2" my="$1" />
@@ -206,10 +146,8 @@ function WalletEditButtonView({
       wallet,
       showBackupButton,
       showDeviceManagementButton,
-      canBatchCreateAccount,
       showRemoveDeviceButton,
       showRemoveWalletButton,
-      activeAccount,
     ],
   );
