+++ packages/kit/src/views/AccountManagerStacks/pages/AccountSelectorStack/WalletDetails/WalletDetailsHeader/AccountSearchBar.tsx
@@ -2,21 +2,57 @@ import { useIntl } from 'react-intl';
 import { StyleSheet } from 'react-native';
 import { useDebouncedCallback } from 'use-debounce';
 
-import { InputUnControlled, XStack } from '@onekeyhq/components';
+import {
+  ActionList,
+  IconButton,
+  InputUnControlled,
+  XStack,
+} from '@onekeyhq/components';
+import { useActiveAccount } from '@onekeyhq/kit/src/states/jotai/contexts/accountSelector';
+import { BatchCreateAccountButton } from '@onekeyhq/kit/src/views/AccountManagerStacks/components/WalletEdit/BatchCreateAccountButton';
+import type {
+  IDBDevice,
+  IDBWallet,
+} from '@onekeyhq/kit-bg/src/dbs/local/types';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
+import accountUtils from '@onekeyhq/shared/src/utils/accountUtils';
+
+import { useAddAccount } from '../hooks/useAddAccount';
 
 export function AccountSearchBar({
   searchText,
   onSearchTextChange,
+  num,
+  isOthersUniversal,
+  focusedWalletInfo,
 }: {
   searchText: string;
   onSearchTextChange: (text: string) => void;
+  num: number;
+  isOthersUniversal: boolean;
+  focusedWalletInfo:
+    | {
+        wallet: IDBWallet;
+        device: IDBDevice | undefined;
+      }
+    | undefined;
 }) {
   const intl = useIntl();
+  const { activeAccount } = useActiveAccount({ num });
+  const { handleAddAccount } = useAddAccount({
+    num,
+    isOthersUniversal,
+    focusedWalletInfo,
+  });
 
   const handleSearch = useDebouncedCallback((text: string) => {
     onSearchTextChange(text?.trim() || '');
   }, 300);
+
+  // Check if bulk create account is available
+  const canBatchCreateAccount =
+    focusedWalletInfo?.wallet?.id &&
+    !accountUtils.isQrWallet({ walletId: focusedWalletInfo?.wallet?.id });
 
   return (
     <XStack
@@ -25,6 +61,8 @@ export function AccountSearchBar({
       py="$2"
       borderBottomWidth={StyleSheet.hairlineWidth}
       borderBottomColor="$neutral3"
+      alignItems="center"
+      gap="$2"
     >
       <InputUnControlled
         leftIconName="SearchOutline"
@@ -41,6 +79,40 @@ export function AccountSearchBar({
         }}
         defaultValue={searchText}
         onChangeText={handleSearch}
+      />
+
+      <ActionList
+        title={intl.formatMessage({ id: ETranslations.global_add_account })}
+        renderTrigger={
+          <IconButton
+            testID="account-search-bar-add-button"
+            icon="PlusSmallOutline"
+            size="small"
+          />
+        }
+        renderItems={({ handleActionListClose }) => (
+          <>
+            <ActionList.Item
+              testID="add-account-button"
+              icon="PlusSmallOutline"
+              label={intl.formatMessage({
+                id: ETranslations.global_add_account,
+              })}
+              onClose={handleActionListClose}
+              onPress={() => {
+                void handleAddAccount();
+                handleActionListClose();
+              }}
+            />
+            {canBatchCreateAccount ? (
+              <BatchCreateAccountButton
+                focusedWalletInfo={focusedWalletInfo}
+                activeAccount={activeAccount}
+                onClose={handleActionListClose}
+              />
+            ) : null}
+          </>
+        )}
       />
     </XStack>
   );
