+++ packages/shared/src/locale/enum/translations.ts
@@ -1268,6 +1268,7 @@
   global_liquidity = 'global.liquidity',
   global_logout = 'global.logout',
   global_manage = 'global.manage',
+  global_manage_accounts = 'global.manage_accounts',
   global_mark_all_as_confirmation_desc = 'global.mark_all_as_confirmation_desc',
   global_mark_all_as_confirmation_title = 'global.mark_all_as_confirmation_title',
   global_market = 'global.market',
@@ -2115,8 +2116,12 @@
   receive_token_list_footer_text = 'receive_token_list_footer_text',
   referral_address_update_desc = 'referral.address_update_desc',
   referral_address_updated = 'referral.address_updated',
+  referral_cumulative_rewards = 'referral.cumulative_rewards',
+  referral_distributed = 'referral.distributed',
   referral_earn_reward = 'referral.earn_reward',
   referral_earn_reward_desc = 'referral.earn_reward_desc',
+  referral_earn_reward_details = 'referral.earn_reward_details',
+  referral_earn_reward_details_desc = 'referral.earn_reward_details_desc',
   referral_earn_reward_tips = 'referral.earn_reward_tips',
   referral_faqs = 'referral.faqs',
   referral_faqs_a1 = 'referral.faqs_a1',
@@ -2130,7 +2135,9 @@
   referral_how_2 = 'referral.how_2',
   referral_how_3 = 'referral.how_3',
   referral_how_4 = 'referral.how_4',
+  referral_hw_level_rebate_rate = 'referral.hw_level_rebate_rate',
   referral_hw_level_title = 'referral.hw_level_title',
+  referral_hw_level_up_remain = 'referral.hw_level_up_remain',
   referral_hw_sales_title = 'referral.hw_sales_title',
   referral_intro_for_you = 'referral.intro_for_you',
   referral_intro_for_you_1 = 'referral.intro_for_you_1',
@@ -2143,6 +2150,8 @@
   referral_intro_title_2 = 'referral.intro_title_2',
   referral_invalid_code = 'referral.invalid_code',
   referral_more_questions = 'referral.more_questions',
+  referral_next_distribution = 'referral.next_distribution',
+  referral_next_distribution_day = 'referral.next_distribution_day',
   referral_no_reward = 'referral.no_reward',
   referral_onboard_bind_code = 'referral.onboard_bind_code',
   referral_onboard_bind_code_finish = 'referral.onboard_bind_code_finish',
@@ -2189,6 +2198,7 @@
   referral_total_reward = 'referral.total_reward',
   referral_total_reward_pop1 = 'referral.total_reward_pop1',
   referral_total_reward_pop2 = 'referral.total_reward_pop2',
+  referral_undistributed = 'referral.undistributed',
   referral_wallet_code_desc = 'referral.wallet_code_desc',
   referral_wallet_code_placeholder = 'referral.wallet_code_placeholder',
   referral_wallet_code_title = 'referral.wallet_code_title',
