+++ packages/shared/src/locale/json/bn.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "নিখুঁতি",
   "global.logout": "লগ আউট",
   "global.manage": "পরিচালনা করুন",
+  "global.manage_accounts": "অ্যাকাউন্ট পরিচালনা করুন",
   "global.mark_all_as_confirmation_desc": "এটি পূর্বাবস্থায় ফেরানো যাবে না।",
   "global.mark_all_as_confirmation_title": "সব পড়া হিসেবে চিহ্নিত করবেন?",
   "global.market": "বাজার",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "পুনরুদ্ধার বাক্যাংশ",
   "global.refresh": "রিফ্রেশ",
   "global.remove": "অপসারণ করুন",
-  "global.remove_account_desc": "আপনি পরে এই ওয়ালেটে \"Bulk add accounts\" ব্যবহার করে এই অ্যাকাউন্টটি পুনরুদ্ধার করতে পারবেন।",
+  "global.remove_account_desc": "আপনি পরে এই ওয়ালেটে \"অ্যাকাউন্ট ম্যানেজ করুন\" ব্যবহার করে এই অ্যাকাউন্টটি পুনরুদ্ধার করতে পারবেন।",
   "global.remove_account_name": "{account} মুছে ফেলুন",
   "global.rename": "নাম পরিবর্তন করুন",
   "global.reset": "রিসেট",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "টোকেন খুঁজে পাচ্ছেন না? খোঁজার চেষ্টা করুন",
   "referral.address_update_desc": "একটি কোড {mail} এ পাঠানো হয়েছে, ঠিকানা আপডেট করার আগে দয়া করে যাচাই করুন।",
   "referral.address_updated": "প্রাপ্তির ঠিকানা আপডেট করা হয়েছে",
+  "referral.cumulative_rewards": "সমষ্টিগত পুরস্কার",
+  "referral.distributed": "বিতরণকৃত",
   "referral.earn_reward": "পুরস্কার অর্জন করুন",
   "referral.earn_reward_desc": "আপনার বন্ধুদের থেকে কমিশন পান, পুরস্কার অর্জন করুন।",
+  "referral.earn_reward_details": "পুরস্কারের বিবরণ",
+  "referral.earn_reward_details_desc": "প্রতি মাসের ১ তারিখে সব টোকেন মার্কেট মূল্যে USDC-তে স্ন্যাপশট নেওয়া হয়, তারপর বিতরণ করা হয়।",
   "referral.earn_reward_tips": "পুরস্কারগুলি আপনার ওয়ালেটে আগামী মাসের ১০ তারিখের মধ্যে বিতরণ করা হবে।",
   "referral.faqs": "প্রশ্নোত্তর",
   "referral.faqs_a1": "আপনার রেফারেল কোড ব্যবহার করার জন্য বন্ধুদের আমন্ত্রণ জানান Earn বা হার্ডওয়্যার ওয়ালেট কেনাকাটার জন্য পুরস্কার পেতে। নিষ্পত্তির প্রায় এক মাস পর আপনার পুরস্কার ঠিকানায় পুরস্কার পাঠানো হবে।",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "আপনার রেফারেল কোড দিয়ে আপনার বন্ধুকে আমন্ত্রণ জানান",
   "referral.how_3": "তারা হার্ডওয়্যার ওয়ালেট কেনে বা আপনার কোড দিয়ে ওয়ালেট যুক্ত করে",
   "referral.how_4": "আপনার পুরস্কার পান",
+  "referral.hw_level_rebate_rate": "{percent} পুরস্কার",
   "referral.hw_level_title": "লেভেল",
+  "referral.hw_level_up_remain": "{LevelName} স্তরে যেতে আরও {Amount} প্রয়োজন",
   "referral.hw_sales_title": "মাসিক বিক্রয়",
   "referral.intro_for_you": "আপনার জন্য",
   "referral.intro_for_you_1": "{RebateRate} পর্যন্ত হার্ডওয়্যার ওয়ালেট বিক্রয় কমিশন পান",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "এটি কীভাবে কাজ করে",
   "referral.invalid_code": "অবৈধ রেফারেল কোড",
   "referral.more_questions": "আরও প্রশ্নের জন্য, <url>https://help.onekey.so/hc/articles/12278189850127<underline>এখানে</underline></url> দেখুন।",
+  "referral.next_distribution": "পরবর্তী বিতরণ",
+  "referral.next_distribution_day": "{Month} ১০ তারিখের মধ্যে",
   "referral.no_reward": "এখনও কোন পুরস্কার নেই",
   "referral.onboard_bind_code": "রেফারেল কোড সংযুক্ত করুন",
   "referral.onboard_bind_code_finish": "শেষ",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "মোট পুরস্কার",
   "referral.total_reward_pop1": "সমস্ত আমন্ত্রণের জন্য মোট নিশ্চিত পুরস্কার",
   "referral.total_reward_pop2": "অমীমাংসিত পুরস্কার অন্তর্ভুক্ত নয়",
+  "referral.undistributed": "অবিতরণকৃত",
   "referral.wallet_code_desc": "একবার বাঁধা হয়ে গেলে, এই ওয়ালেট থেকে প্রাপ্ত সমস্ত ঠিকানা এই রেফারাল কোডের সাথে যুক্ত হবে এবং পরিবর্তন করা যাবে না।",
   "referral.wallet_code_placeholder": "রেফারেল কোড",
   "referral.wallet_code_title": "আপনার কি কোনো রেফারেল কোড আছে?",
