+++ packages/shared/src/locale/json/de.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Liquidität",
   "global.logout": "Abmelden",
   "global.manage": "Verwalten",
+  "global.manage_accounts": "Konten verwalten",
   "global.mark_all_as_confirmation_desc": "Dies kann nicht rückgängig gemacht werden.",
   "global.mark_all_as_confirmation_title": "Alle als gelesen markieren?",
   "global.market": "Markt",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Wiederherstellungsphrase",
   "global.refresh": "Aktualisieren",
   "global.remove": "Entfernen",
-  "global.remove_account_desc": "Sie können dieses Konto später in dieser Wallet wiederherstellen, indem Sie \"Konten gesammelt hinzufügen\" verwenden.",
+  "global.remove_account_desc": "Sie können dieses Konto später in dieser Wallet wiederherstellen, indem Sie \"Konten verwalten\" verwenden.",
   "global.remove_account_name": "Entfernen Sie {account}",
   "global.rename": "Umbenennen",
   "global.reset": "Zurücksetzen",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Token nicht gefunden? Suche weiter",
   "referral.address_update_desc": "Ein Code wurde an {mail} gesendet, bitte verifizieren Sie ihn, bevor Sie die Adresse aktualisieren.",
   "referral.address_updated": "Empfangsadresse aktualisiert",
+  "referral.cumulative_rewards": "Kumulative Belohnungen",
+  "referral.distributed": "Verteilt",
   "referral.earn_reward": "Belohnung verdienen",
   "referral.earn_reward_desc": "Erhalten Sie Provision von Ihren Freunden und verdienen Sie Belohnungen.",
+  "referral.earn_reward_details": "Belohnungsdetails",
+  "referral.earn_reward_details_desc": "Alle Token werden am 1. eines jeden Monats zum Marktwert in USDC festgehalten und anschließend verteilt.",
   "referral.earn_reward_tips": "Belohnungen werden bis zum 10. des nächsten Monats an Ihre Wallet verteilt.",
   "referral.faqs": "Häufig gestellte Fragen",
   "referral.faqs_a1": "Laden Sie Freunde ein, Ihren Empfehlungscode für Earn- oder Hardware-Wallet-Käufe zu verwenden, um Belohnungen zu erhalten. Belohnungen werden etwa einen Monat nach der Abwicklung an Ihre Belohnungsadresse gesendet.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Laden Sie Ihren Freund mit Ihrem Empfehlungscode ein",
   "referral.how_3": "Sie kaufen eine Hardware-Wallet oder verbinden eine Wallet mit Ihrem Code.",
   "referral.how_4": "Ihr erhaltet Belohnung",
+  "referral.hw_level_rebate_rate": "{percent} Belohnung",
   "referral.hw_level_title": "Stufe",
+  "referral.hw_level_up_remain": "{Amount} mehr bis zum Level {LevelName}",
   "referral.hw_sales_title": "Monatlicher Umsatz",
   "referral.intro_for_you": "Für Sie",
   "referral.intro_for_you_1": "Erhalten Sie bis zu {RebateRate} Verkaufsprovision für Hardware-Wallets",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Wie es funktioniert",
   "referral.invalid_code": "Ungültiger Empfehlungscode",
   "referral.more_questions": "Für weitere Fragen, schauen Sie <url>https://help.onekey.so/hc/articles/12278189850127<underline>hier</underline></url> nach.",
+  "referral.next_distribution": "Nächste Verteilung",
+  "referral.next_distribution_day": "Bis {Month} 10",
   "referral.no_reward": "Noch keine Belohnung",
   "referral.onboard_bind_code": "Empfehlungscode verknüpfen",
   "referral.onboard_bind_code_finish": "Beenden",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Gesamte Belohnung",
   "referral.total_reward_pop1": "Gesamt bestätigte Belohnungen für alle Einladungen",
   "referral.total_reward_pop2": "Ausstehende Belohnungen sind nicht enthalten",
+  "referral.undistributed": "Nicht verteilt",
   "referral.wallet_code_desc": "Sobald gebunden, werden alle von dieser Wallet abgeleiteten Adressen mit diesem Empfehlungscode verknüpft und können nicht mehr geändert werden.",
   "referral.wallet_code_placeholder": "Empfehlungscode",
   "referral.wallet_code_title": "Haben Sie einen Empfehlungscode?",
