+++ packages/shared/src/locale/json/en_US.json
@@ -1048,7 +1048,7 @@
   "global.bulk_add_account_dnx_error": "Bulk add accounts error, please connect OneKey Classic",
   "global.bulk_add_accounts": "Bulk add accounts",
   "global.bulk_add_accounts_error": "You can add a maximum of {amount} accounts",
-  "global.bulk_copy_addresses": "Bulk Copy Addresses",
+  "global.bulk_copy_addresses": "Bulk copy addresses",
   "global.bulk_copy_addresses_addresses_copied": "Addresses copied",
   "global.bulk_copy_addresses_export_csv": "Export CSV",
   "global.bulk_copy_addresses_tabs_my_accounts": "My accounts",
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Liquidity",
   "global.logout": "Logout",
   "global.manage": "Manage",
+  "global.manage_accounts": "Manage accounts",
   "global.mark_all_as_confirmation_desc": "This can’t be undone.",
   "global.mark_all_as_confirmation_title": "Mark all as read?",
   "global.market": "Market",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Recovery phrase",
   "global.refresh": "Refresh",
   "global.remove": "Remove",
-  "global.remove_account_desc": "You can restore this account later in this wallet by using \"Bulk add accounts\".",
+  "global.remove_account_desc": "You can restore this account later in this wallet by using \"Manage accounts\".",
   "global.remove_account_name": "Remove {account}",
   "global.rename": "Rename",
   "global.reset": "Reset",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Can't find token? Try searching",
   "referral.address_update_desc": "A code has been sent to {mail}, please verify before updating the address.",
   "referral.address_updated": "Receiving address updated",
+  "referral.cumulative_rewards": "Cumulative rewards",
+  "referral.distributed": "Distributed",
   "referral.earn_reward": "Earn reward",
   "referral.earn_reward_desc": "Get commission from your friends’ Earn rewards.",
+  "referral.earn_reward_details": "Reward details",
+  "referral.earn_reward_details_desc": "All tokens are snapshotted to USDC at market value on the 1st of each month, then distributed.",
   "referral.earn_reward_tips": "Rewards will be distributed to your wallet by the 10th of next month.",
   "referral.faqs": "FAQs",
   "referral.faqs_a1": "Invite friends to use your referral code for Earn or hardware wallet purchases to get rewards. Rewards will be sent to your reward address about a month after settlement.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Invite your friend by your referral code",
   "referral.how_3": "They buy hardware wallet or bind wallet with your code",
   "referral.how_4": "Your get reward",
+  "referral.hw_level_rebate_rate": "{percent} reward",
   "referral.hw_level_title": "Level",
+  "referral.hw_level_up_remain": "{Amount} more to {LevelName} level",
   "referral.hw_sales_title": "Monthly sales",
   "referral.intro_for_you": "For You",
   "referral.intro_for_you_1": "Get up to {RebateRate} hardware wallet sale commission",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "How it works",
   "referral.invalid_code": "Invalid referral code",
   "referral.more_questions": "For more questions, check <url>https://help.onekey.so/hc/articles/12278189850127<underline>here</underline></url>.",
+  "referral.next_distribution": "Next distribution",
+  "referral.next_distribution_day": "By {Month} 10",
   "referral.no_reward": "No reward yet",
   "referral.onboard_bind_code": "Bind referral code",
   "referral.onboard_bind_code_finish": "Finish",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Total Reward",
   "referral.total_reward_pop1": "Total confirmed rewards for all referrals",
   "referral.total_reward_pop2": "Pending rewards are not included",
+  "referral.undistributed": "Undistributed",
   "referral.wallet_code_desc": "Once bound, all addresses derived from this wallet will be associated with this referral code and cannot be changed.",
   "referral.wallet_code_placeholder": "Referral code",
   "referral.wallet_code_title": "Do you have a referral code?",
