+++ packages/shared/src/locale/json/es.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Liquidez",
   "global.logout": "Cerrar sesión",
   "global.manage": "Administrar",
+  "global.manage_accounts": "Administrar cuentas",
   "global.mark_all_as_confirmation_desc": "Esto no se puede deshacer.",
   "global.mark_all_as_confirmation_title": "¿Marcar todo como leído?",
   "global.market": "Mercado",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Frase de recuperación",
   "global.refresh": "Actualizar",
   "global.remove": "Eliminar",
-  "global.remove_account_desc": "Puedes restaurar esta cuenta más tarde en esta billetera usando \"Agregar cuentas en bloque\".",
+  "global.remove_account_desc": "Puedes restaurar esta cuenta más tarde en esta billetera usando \"Administrar cuentas\".",
   "global.remove_account_name": "Eliminar {account}",
   "global.rename": "Renombrar",
   "global.reset": "Resetear",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "¿No encuentras el token? Intenta buscarlo",
   "referral.address_update_desc": "Se ha enviado un código a {mail}, por favor verifique antes de actualizar la dirección.",
   "referral.address_updated": "Dirección de recepción actualizada",
+  "referral.cumulative_rewards": "Recompensas acumulativas",
+  "referral.distributed": "Distribuido",
   "referral.earn_reward": "Gana recompensa",
   "referral.earn_reward_desc": "Obtén comisión de las ganancias de tus amigos. Gana recompensas.",
+  "referral.earn_reward_details": "Detalles de la recompensa",
+  "referral.earn_reward_details_desc": "Todos los tokens se convierten a USDC a valor de mercado el día 1 de cada mes y luego se distribuyen.",
   "referral.earn_reward_tips": "Las recompensas se distribuirán a su billetera antes del 10 del próximo mes.",
   "referral.faqs": "Preguntas frecuentes",
   "referral.faqs_a1": "Invita a tus amigos a usar tu código de referencia para compras de Earn o de billeteras de hardware para obtener recompensas. Las recompensas se enviarán a tu dirección de recompensas aproximadamente un mes después de la liquidación.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Invita a tu amigo con tu código de referencia",
   "referral.how_3": "Compran una billetera de hardware o vinculan una billetera con tu código",
   "referral.how_4": "Tu obtienes recompensa",
+  "referral.hw_level_rebate_rate": "Recompensa del {percent}",
   "referral.hw_level_title": "Nivel",
+  "referral.hw_level_up_remain": "{Amount} más para alcanzar el nivel {LevelName}",
   "referral.hw_sales_title": "Ventas mensuales",
   "referral.intro_for_you": "Para ti",
   "referral.intro_for_you_1": "Obtén hasta {RebateRate} de comisión por venta de monederos de hardware",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Cómo funciona",
   "referral.invalid_code": "Código de referencia no válido",
   "referral.more_questions": "Para más preguntas, consulte <url>https://help.onekey.so/hc/articles/12278189850127<underline>aquí</underline></url>.",
+  "referral.next_distribution": "Próxima distribución",
+  "referral.next_distribution_day": "Para el {Month} 10",
   "referral.no_reward": "Aún no hay recompensa",
   "referral.onboard_bind_code": "Vincular código de referido",
   "referral.onboard_bind_code_finish": "Finalizar",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Recompensa Total",
   "referral.total_reward_pop1": "Recompensas totales confirmadas por todas las invitaciones",
   "referral.total_reward_pop2": "Las recompensas pendientes no están incluidas",
+  "referral.undistributed": "Sin distribuir",
   "referral.wallet_code_desc": "Una vez vinculado, todas las direcciones derivadas de esta cartera estarán asociadas con este código de referencia y no se podrán cambiar.",
   "referral.wallet_code_placeholder": "Código de referencia",
   "referral.wallet_code_title": "¿Tienes un código de referencia?",
