+++ packages/shared/src/locale/json/fr_FR.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Liquidité",
   "global.logout": "Se déconnecter",
   "global.manage": "Gérer",
+  "global.manage_accounts": "Gérer les comptes",
   "global.mark_all_as_confirmation_desc": "Cela ne peut pas être annulé.",
   "global.mark_all_as_confirmation_title": "Marquer tout comme lu ?",
   "global.market": "Marché",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Phrase de récupération",
   "global.refresh": "Actualiser",
   "global.remove": "Supprimer",
-  "global.remove_account_desc": "Vous pourrez restaurer ce compte ultérieurement dans ce portefeuille en utilisant \\\"Ajouter des comptes en masse\\\".",
+  "global.remove_account_desc": "Vous pourrez restaurer ce compte ultérieurement dans ce portefeuille en utilisant \\\"Gérer les comptes\\\".",
   "global.remove_account_name": "Supprimer {account}",
   "global.rename": "Renommer",
   "global.reset": "Réinitialiser",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Vous ne trouvez pas de jeton ? Essayez de rechercher",
   "referral.address_update_desc": "Un code a été envoyé à {mail}, veuillez vérifier avant de mettre à jour l'adresse.",
   "referral.address_updated": "Adresse de réception mise à jour",
+  "referral.cumulative_rewards": "Récompenses cumulées",
+  "referral.distributed": "Distribué",
   "referral.earn_reward": "Gagnez une récompense",
   "referral.earn_reward_desc": "Obtenez une commission sur les récompenses gagnées par vos amis.",
+  "referral.earn_reward_details": "Détails de la récompense",
+  "referral.earn_reward_details_desc": "Tous les jetons sont convertis en USDC à leur valeur marchande le 1er de chaque mois, puis distribués.",
   "referral.earn_reward_tips": "Les récompenses seront distribuées dans votre portefeuille d'ici le 10 du mois prochain.",
   "referral.faqs": "FAQ",
   "referral.faqs_a1": "Invitez des amis à utiliser votre code de parrainage pour Earn ou pour l'achat de portefeuilles matériels afin d'obtenir des récompenses. Les récompenses seront envoyées à votre adresse de récompense environ un mois après le règlement.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Invitez votre ami avec votre code de parrainage",
   "referral.how_3": "Ils achètent un portefeuille matériel ou associent un portefeuille avec votre code",
   "referral.how_4": "Vous obtenez une récompense",
+  "referral.hw_level_rebate_rate": "{percent} de récompense",
   "referral.hw_level_title": "Niveau",
+  "referral.hw_level_up_remain": "{Amount} de plus pour atteindre le niveau {LevelName}",
   "referral.hw_sales_title": "Ventes mensuelles",
   "referral.intro_for_you": "Pour vous",
   "referral.intro_for_you_1": "Obtenez jusqu'à {RebateRate} de commission sur la vente de portefeuilles matériels",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Comment ça fonctionne",
   "referral.invalid_code": "Code de parrainage invalide",
   "referral.more_questions": "Pour plus de questions, consultez <url>https://help.onekey.so/hc/articles/12278189850127<underline>ici</underline></url>.",
+  "referral.next_distribution": "Prochaine distribution",
+  "referral.next_distribution_day": "D'ici le 10 {Month}",
   "referral.no_reward": "Pas encore de récompense",
   "referral.onboard_bind_code": "Lier le code de parrainage",
   "referral.onboard_bind_code_finish": "Finition",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Récompense Totale",
   "referral.total_reward_pop1": "Récompenses totales confirmées pour toutes les invitations",
   "referral.total_reward_pop2": "Les récompenses en attente ne sont pas incluses",
+  "referral.undistributed": "Non distribué",
   "referral.wallet_code_desc": "Une fois lié, toutes les adresses dérivées de ce portefeuille seront associées à ce code de parrainage et ne pourront pas être modifiées.",
   "referral.wallet_code_placeholder": "Code de parrainage",
   "referral.wallet_code_title": "Avez-vous un code de parrainage ?",
