+++ packages/shared/src/locale/json/hi_IN.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "तरलता",
   "global.logout": "लॉग आउट",
   "global.manage": "प्रबंधन करें",
+  "global.manage_accounts": "खातों का प्रबंधन करें",
   "global.mark_all_as_confirmation_desc": "इसे पूर्ववत नहीं किया जा सकता.",
   "global.mark_all_as_confirmation_title": "सभी को पढ़ा हुआ मार्क करें?",
   "global.market": "बाजार",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "रिकवरी वाक्यांश",
   "global.refresh": "ताजा करें",
   "global.remove": "हटाएं",
-  "global.remove_account_desc": "आप इस खाते को बाद में इस वॉलेट में \"Bulk add accounts\" का उपयोग करके पुनर्स्थापित कर सकते हैं।",
+  "global.remove_account_desc": "आप इस खाते को बाद में इस वॉलेट में \"खाते प्रबंधित करें\" का उपयोग करके पुनर्स्थापित कर सकते हैं।",
   "global.remove_account_name": "{account} को हटाएं",
   "global.rename": "नाम बदलें",
   "global.reset": "रीसेट",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "टोकन नहीं मिल रहा? खोजने का प्रयास करें",
   "referral.address_update_desc": "{mail} पर एक कोड भेजा गया है, कृपया पता अपडेट करने से पहले सत्यापित करें।",
   "referral.address_updated": "प्राप्ति पता अपडेट किया गया",
+  "referral.cumulative_rewards": "संचयी पुरस्कार",
+  "referral.distributed": "वितरित",
   "referral.earn_reward": "इनाम अर्जित करें",
   "referral.earn_reward_desc": "अपने दोस्तों से कमीशन प्राप्त करें, पुरस्कार अर्जित करें।",
+  "referral.earn_reward_details": "इनाम विवरण",
+  "referral.earn_reward_details_desc": "सभी टोकन को हर महीने की पहली तारीख को बाजार मूल्य पर USDC में स्नैपशॉट किया जाता है, फिर वितरित किया जाता है।",
   "referral.earn_reward_tips": "पुरस्कार अगले महीने की 10 तारीख तक आपके वॉलेट में वितरित कर दिए जाएंगे।",
   "referral.faqs": "सामान्य प्रश्न",
   "referral.faqs_a1": "अपने दोस्तों को अपने रेफरल कोड का उपयोग करके अर्न या हार्डवेयर वॉलेट खरीदने के लिए आमंत्रित करें ताकि आपको इनाम मिल सके। इनाम आपके इनाम पते पर निपटान के लगभग एक महीने बाद भेजा जाएगा।",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "अपने मित्र को अपने रेफरल कोड द्वारा आमंत्रित करें",
   "referral.how_3": "वे हार्डवेयर वॉलेट खरीदते हैं या आपके कोड के साथ वॉलेट को जोड़ते हैं",
   "referral.how_4": "आपको पुरस्कार प्राप्त होता है",
+  "referral.hw_level_rebate_rate": "{percent} इनाम",
   "referral.hw_level_title": "स्तर",
+  "referral.hw_level_up_remain": "{LevelName} स्तर तक पहुँचने के लिए {Amount} और चाहिए",
   "referral.hw_sales_title": "मासिक बिक्री",
   "referral.intro_for_you": "आपके लिए",
   "referral.intro_for_you_1": "{RebateRate} हार्डवेयर वॉलेट बिक्री कमीशन तक प्राप्त करें",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "यह कैसे काम करता है",
   "referral.invalid_code": "अमान्य रेफ़रल कोड",
   "referral.more_questions": "अधिक प्रश्नों के लिए, <url>https://help.onekey.so/hc/articles/12278189850127<underline>यहां</underline></url> जांचें।",
+  "referral.next_distribution": "अगला वितरण",
+  "referral.next_distribution_day": "{Month} 10 तक",
   "referral.no_reward": "अभी तक कोई इनाम नहीं",
   "referral.onboard_bind_code": "रेफरल कोड जोड़ें",
   "referral.onboard_bind_code_finish": "खत्म करना",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "कुल इनाम",
   "referral.total_reward_pop1": "सभी निमंत्रणों के लिए कुल पुष्टि की गई पुरस्कार",
   "referral.total_reward_pop2": "लंबित पुरस्कार शामिल नहीं हैं",
+  "referral.undistributed": "अवितरित",
   "referral.wallet_code_desc": "एक बार बाउंड होने के बाद, इस वॉलेट से प्राप्त सभी पते इस रेफरल कोड से जुड़े होंगे और इन्हें बदला नहीं जा सकता है।",
   "referral.wallet_code_placeholder": "रेफरल कोड",
   "referral.wallet_code_title": "क्या आपके पास कोई रेफरल कोड है?",
