+++ packages/shared/src/locale/json/id.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Likuiditas",
   "global.logout": "Keluar",
   "global.manage": "Kelola",
+  "global.manage_accounts": "Kelola akun",
   "global.mark_all_as_confirmation_desc": "Ini tidak dapat dibatalkan.",
   "global.mark_all_as_confirmation_title": "Tandai semua sebagai telah dibaca?",
   "global.market": "Pasar",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Frase pemulihan",
   "global.refresh": "Segarkan",
   "global.remove": "Hapus",
-  "global.remove_account_desc": "Anda dapat memulihkan akun ini nanti di dompet ini dengan menggunakan \"Tambah akun massal\".",
+  "global.remove_account_desc": "Anda dapat memulihkan akun ini nanti di dompet ini dengan menggunakan \"Kelola akun\".",
   "global.remove_account_name": "Hapus {account}",
   "global.rename": "Ganti Nama",
   "global.reset": "Mengatur ulang",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Tidak dapat menemukan token? Coba cari",
   "referral.address_update_desc": "Kode telah dikirim ke {mail}, silakan verifikasi sebelum memperbarui alamat.",
   "referral.address_updated": "Alamat penerima diperbarui",
+  "referral.cumulative_rewards": "Hadiah kumulatif",
+  "referral.distributed": "Terdistribusi",
   "referral.earn_reward": "Dapatkan hadiah",
   "referral.earn_reward_desc": "Dapatkan komisi dari teman-teman Anda, Dapatkan hadiah.",
+  "referral.earn_reward_details": "Detail hadiah",
+  "referral.earn_reward_details_desc": "Semua token di-snapshot ke USDC berdasarkan nilai pasar pada tanggal 1 setiap bulan, lalu didistribusikan.",
   "referral.earn_reward_tips": "Hadiah akan didistribusikan ke dompet Anda pada tanggal 10 bulan depan.",
   "referral.faqs": "Pertanyaan yang Sering Diajukan",
   "referral.faqs_a1": "Undang teman untuk menggunakan kode referal Anda untuk pembelian Earn atau dompet perangkat keras untuk mendapatkan hadiah. Hadiah akan dikirim ke alamat hadiah Anda sekitar sebulan setelah penyelesaian.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Undang teman Anda dengan kode referensi Anda",
   "referral.how_3": "Mereka membeli dompet perangkat keras atau menghubungkan dompet dengan kode Anda",
   "referral.how_4": "Anda mendapatkan hadiah",
+  "referral.hw_level_rebate_rate": "Hadiah {percent}",
   "referral.hw_level_title": "Tingkat",
+  "referral.hw_level_up_remain": "{Amount} lagi untuk mencapai level {LevelName}",
   "referral.hw_sales_title": "Penjualan bulanan",
   "referral.intro_for_you": "Untuk Anda",
   "referral.intro_for_you_1": "Dapatkan komisi penjualan dompet perangkat keras hingga {RebateRate}",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Cara kerjanya",
   "referral.invalid_code": "Kode rujukan tidak valid",
   "referral.more_questions": "Untuk pertanyaan lebih lanjut, periksa <url>https://help.onekey.so/hc/articles/12278189850127<underline>di sini</underline></url>.",
+  "referral.next_distribution": "Distribusi berikutnya",
+  "referral.next_distribution_day": "Sampai {Month} 10",
   "referral.no_reward": "Belum ada hadiah",
   "referral.onboard_bind_code": "Ikat kode referensi",
   "referral.onboard_bind_code_finish": "Menyelesaikan",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Total Hadiah",
   "referral.total_reward_pop1": "Total hadiah yang dikonfirmasi untuk semua undangan",
   "referral.total_reward_pop2": "Hadiah yang tertunda tidak termasuk",
+  "referral.undistributed": "Belum didistribusikan",
   "referral.wallet_code_desc": "Setelah terikat, semua alamat yang berasal dari dompet ini akan dikaitkan dengan kode referral ini dan tidak dapat diubah.",
   "referral.wallet_code_placeholder": "Kode rujukan",
   "referral.wallet_code_title": "Apakah Anda memiliki kode referral?",
