+++ packages/shared/src/locale/json/it_IT.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Liquidità",
   "global.logout": "Esci",
   "global.manage": "Gestisci",
+  "global.manage_accounts": "Gestisci account",
   "global.mark_all_as_confirmation_desc": "Questa operazione non può essere annullata.",
   "global.mark_all_as_confirmation_title": "Contrassegnare tutto come letto?",
   "global.market": "Mercato",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Frase di recupero",
   "global.refresh": "Aggiorna",
   "global.remove": "Rimuovi",
-  "global.remove_account_desc": "Puoi ripristinare questo account in seguito in questo wallet utilizzando \"Aggiunta multipla di account\".",
+  "global.remove_account_desc": "Puoi ripristinare questo account in seguito in questo portafoglio utilizzando \"Gestisci account\".",
   "global.remove_account_name": "Rimuovi {account}",
   "global.rename": "Rinomina",
   "global.reset": "Resettare",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Non riesci a trovare il token? Prova a cercare",
   "referral.address_update_desc": "Un codice è stato inviato a {mail}, si prega di verificare prima di aggiornare l'indirizzo.",
   "referral.address_updated": "Indirizzo di ricezione aggiornato",
+  "referral.cumulative_rewards": "Ricompense cumulative",
+  "referral.distributed": "Distribuito",
   "referral.earn_reward": "Guadagna ricompensa",
   "referral.earn_reward_desc": "Ottieni commissioni dai tuoi amici e guadagna ricompense.",
+  "referral.earn_reward_details": "Dettagli della ricompensa",
+  "referral.earn_reward_details_desc": "Tutti i token vengono convertiti in USDC al valore di mercato il primo di ogni mese, quindi distribuiti.",
   "referral.earn_reward_tips": "Le ricompense saranno distribuite al tuo portafoglio entro il 10 del mese prossimo.",
   "referral.faqs": "Domande frequenti",
   "referral.faqs_a1": "Invita gli amici a utilizzare il tuo codice di riferimento per acquisti Earn o di portafogli hardware per ottenere ricompense. Le ricompense verranno inviate al tuo indirizzo di ricompensa circa un mese dopo il regolamento.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Invita il tuo amico con il tuo codice di riferimento",
   "referral.how_3": "Acquistano un hardware wallet o collegano un wallet con il tuo codice",
   "referral.how_4": "Ottieni la tua ricompensa",
+  "referral.hw_level_rebate_rate": "Ricompensa del {percent}",
   "referral.hw_level_title": "Livello",
+  "referral.hw_level_up_remain": "{Amount} ancora per il livello {LevelName}",
   "referral.hw_sales_title": "Vendite mensili",
   "referral.intro_for_you": "Per te",
   "referral.intro_for_you_1": "Ottieni fino a {RebateRate} di commissione sulla vendita di hardware wallet",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Come funziona",
   "referral.invalid_code": "Codice di riferimento non valido",
   "referral.more_questions": "Per ulteriori domande, controlla <url>https://help.onekey.so/hc/articles/12278189850127<underline>qui</underline></url>.",
+  "referral.next_distribution": "Prossima distribuzione",
+  "referral.next_distribution_day": "Entro il {Month} 10",
   "referral.no_reward": "Nessuna ricompensa ancora",
   "referral.onboard_bind_code": "Associa codice di riferimento",
   "referral.onboard_bind_code_finish": "Fine",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Ricompensa Totale",
   "referral.total_reward_pop1": "Totale ricompense confermate per tutti gli inviti",
   "referral.total_reward_pop2": "Le ricompense in sospeso non sono incluse",
+  "referral.undistributed": "Non distribuito",
   "referral.wallet_code_desc": "Una volta associato, tutti gli indirizzi derivati da questo portafoglio saranno associati a questo codice di referral e non potranno essere modificati.",
   "referral.wallet_code_placeholder": "Codice di riferimento",
   "referral.wallet_code_title": "Hai un codice di riferimento?",
