+++ packages/shared/src/locale/json/ja_JP.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "流動性",
   "global.logout": "ログアウト",
   "global.manage": "管理する",
+  "global.manage_accounts": "アカウントを管理する",
   "global.mark_all_as_confirmation_desc": "これは元に戻せません。",
   "global.mark_all_as_confirmation_title": "すべてを既読としてマークしますか?",
   "global.market": "マーケット",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "リカバリーフレーズ",
   "global.refresh": "更新",
   "global.remove": "削除",
-  "global.remove_account_desc": "このアカウントは、「アカウント一括追加」を使って後でこのウォレットに復元できます。",
+  "global.remove_account_desc": "「アカウントの管理」を使って、このウォレットで後からこのアカウントを復元できます。",
   "global.remove_account_name": "{account}を削除",
   "global.rename": "名前を変更する",
   "global.reset": "リセット",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "トークンが見つからない場合は検索してみてください",
   "referral.address_update_desc": "{mail}にコードが送信されました。住所を更新する前に確認してください。",
   "referral.address_updated": "受取先住所が更新されました",
+  "referral.cumulative_rewards": "累積報酬",
+  "referral.distributed": "分散型",
   "referral.earn_reward": "報酬を獲得する",
   "referral.earn_reward_desc": "友達からコミッションを得て、報酬を獲得しましょう。",
+  "referral.earn_reward_details": "報酬の詳細",
+  "referral.earn_reward_details_desc": "すべてのトークンは毎月1日に市場価値でUSDCにスナップショットされ、その後分配されます。",
   "referral.earn_reward_tips": "報酬は来月の10日までにあなたのウォレットに配布されます。",
   "referral.faqs": "よくある質問",
   "referral.faqs_a1": "友達を招待して、Earnまたはハードウェアウォレットの購入にリファラルコードを使用してもらい、報酬を獲得しましょう。報酬は決済後約1か月であなたの報酬アドレスに送信されます。",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "紹介コードで友達を招待してください",
   "referral.how_3": "彼らはハードウェアウォレットを購入するか、あなたのコードでウォレットを紐付けます",
   "referral.how_4": "あなたの報酬を受け取る",
+  "referral.hw_level_rebate_rate": "{percent}の報酬",
   "referral.hw_level_title": "レベル",
+  "referral.hw_level_up_remain": "{LevelName}レベルまであと{Amount}",
   "referral.hw_sales_title": "月次売上",
   "referral.intro_for_you": "あなたのために",
   "referral.intro_for_you_1": "最大{RebateRate}のハードウェアウォレット販売手数料を獲得",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "仕組み",
   "referral.invalid_code": "無効な紹介コード",
   "referral.more_questions": "詳細な質問については、<url>https://help.onekey.so/hc/articles/12278189850127<underline>こちら</underline></url>をご確認ください。",
+  "referral.next_distribution": "次回の配布",
+  "referral.next_distribution_day": "{Month}10日までに",
   "referral.no_reward": "まだ報酬はありません",
   "referral.onboard_bind_code": "紹介コードを紐付ける",
   "referral.onboard_bind_code_finish": "仕上げる",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "総報酬",
   "referral.total_reward_pop1": "すべての招待に対する確認済み報酬の合計",
   "referral.total_reward_pop2": "保留中の報酬は含まれていません",
+  "referral.undistributed": "未配布",
   "referral.wallet_code_desc": "一度紐付けられると、このウォレットから派生したすべてのアドレスはこの紹介コードに関連付けられ、変更できなくなります。",
   "referral.wallet_code_placeholder": "紹介コード",
   "referral.wallet_code_title": "紹介コードはお持ちですか？",
