+++ packages/shared/src/locale/json/ko_KR.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "유동성",
   "global.logout": "로그아웃",
   "global.manage": "관리",
+  "global.manage_accounts": "계정 관리",
   "global.mark_all_as_confirmation_desc": "이는 취소할 수 없습니다.",
   "global.mark_all_as_confirmation_title": "모두 읽음으로 표시하시겠습니까?",
   "global.market": "마켓",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "복구 문구",
   "global.refresh": "새로 고침",
   "global.remove": "제거",
-  "global.remove_account_desc": "\"계정 일괄 추가\"를 사용하여 나중에 이 지갑에서 이 계정을 복원할 수 있습니다.",
+  "global.remove_account_desc": "\"계정 관리\"를 사용하여 나중에 이 지갑에서 이 계정을 복원할 수 있습니다.",
   "global.remove_account_name": "{account} 제거",
   "global.rename": "이름 변경",
   "global.reset": "초기화",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "토큰을 찾을 수 없습니까? 검색해 보세요",
   "referral.address_update_desc": "코드가 {mail}로 전송되었습니다. 주소를 업데이트하기 전에 확인해 주세요.",
   "referral.address_updated": "수신 주소가 업데이트되었습니다",
+  "referral.cumulative_rewards": "누적 보상",
+  "referral.distributed": "분산",
   "referral.earn_reward": "보상 받기",
   "referral.earn_reward_desc": "친구로부터 커미션을 받고 보상을 받으세요.",
+  "referral.earn_reward_details": "보상 세부 정보",
+  "referral.earn_reward_details_desc": "모든 토큰은 매월 1일 시가 기준으로 USDC로 스냅샷이 진행된 후 분배됩니다.",
   "referral.earn_reward_tips": "보상은 다음 달 10일까지 귀하의 지갑으로 배포될 것입니다.",
   "referral.faqs": "자주 묻는 질문",
   "referral.faqs_a1": "친구들을 초대하여 귀하의 추천 코드를 사용하여 Earn 또는 하드웨어 지갑 구매를 통해 보상을 받으세요. 보상은 정산 후 약 한 달 후에 귀하의 보상 주소로 발송됩니다.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "초대코드를 사용해서 친구 초대하기",
   "referral.how_3": "친구가 초대코드를 이용해 하드웨어 지갑을 구매하기",
   "referral.how_4": "보상 수령하기",
+  "referral.hw_level_rebate_rate": "{percent} 보상",
   "referral.hw_level_title": "레벨",
+  "referral.hw_level_up_remain": "{LevelName} 레벨까지 {Amount} 남았습니다",
   "referral.hw_sales_title": "월간 매출",
   "referral.intro_for_you": "추천한 사람",
   "referral.intro_for_you_1": "최대 {RebateRate} 하드웨어 지갑 판매 수수료를 받으세요",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "참여 방법",
   "referral.invalid_code": "잘못된 추천 코드입니다",
   "referral.more_questions": "더 많은 질문은 <url>https://help.onekey.so/hc/articles/12278189850127<underline>여기</underline></url>를 확인하세요.",
+  "referral.next_distribution": "다음 분배",
+  "referral.next_distribution_day": "{Month} 10일까지",
   "referral.no_reward": "아직 보상이 없습니다",
   "referral.onboard_bind_code": "추천 코드 연결",
   "referral.onboard_bind_code_finish": "마치다",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "총 보상",
   "referral.total_reward_pop1": "모든 초대에 대한 총 확인된 보상",
   "referral.total_reward_pop2": "보류 중인 보상은 포함되지 않습니다",
+  "referral.undistributed": "배포되지 않음",
   "referral.wallet_code_desc": "바인딩되면 이 지갑에서 파생된 모든 주소가 이 추천 코드와 연결되며 변경할 수 없습니다.",
   "referral.wallet_code_placeholder": "추천 코드",
   "referral.wallet_code_title": "추천 코드가 있으신가요?",
