+++ packages/shared/src/locale/json/pt.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Liquidez",
   "global.logout": "Sair",
   "global.manage": "Gerenciar",
+  "global.manage_accounts": "Gerenciar contas",
   "global.mark_all_as_confirmation_desc": "Isso não pode ser desfeito.",
   "global.mark_all_as_confirmation_title": "Marcar tudo como lido?",
   "global.market": "Mercado",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Frase de recuperação",
   "global.refresh": "Atualizar",
   "global.remove": "Remover",
-  "global.remove_account_desc": "Você pode restaurar esta conta posteriormente nesta carteira usando \\\"Adicionar contas em massa\\\".",
+  "global.remove_account_desc": "Você pode restaurar esta conta posteriormente nesta carteira usando \\\"Gerenciar contas\\\".",
   "global.remove_account_name": "Remover {account}",
   "global.rename": "Renomear",
   "global.reset": "Redefinir",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Não consegue encontrar o token? Tente pesquisar",
   "referral.address_update_desc": "Um código foi enviado para {mail}, por favor verifique antes de atualizar o endereço.",
   "referral.address_updated": "Endereço de recebimento atualizado",
+  "referral.cumulative_rewards": "Recompensas acumuladas",
+  "referral.distributed": "Distribuído",
   "referral.earn_reward": "Ganhe recompensa",
   "referral.earn_reward_desc": "Ganhe comissão das recompensas Earn dos seus amigos.",
+  "referral.earn_reward_details": "Detalhes da recompensa",
+  "referral.earn_reward_details_desc": "Todos os tokens são convertidos para USDC pelo valor de mercado no dia 1 de cada mês e, em seguida, distribuídos.",
   "referral.earn_reward_tips": "As recompensas serão distribuídas para sua carteira até o dia 10 do próximo mês.",
   "referral.faqs": "Perguntas Frequentes",
   "referral.faqs_a1": "Convide amigos para usar seu código de indicação para compras de Earn ou carteira de hardware para obter recompensas. As recompensas serão enviadas para o seu endereço de recompensas cerca de um mês após a liquidação.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Convide seu amigo usando seu código de referência",
   "referral.how_3": "Eles compram carteira física ou vinculam carteira com seu código",
   "referral.how_4": "Você recebe recompensa",
+  "referral.hw_level_rebate_rate": "{percent} de recompensa",
   "referral.hw_level_title": "Nível",
+  "referral.hw_level_up_remain": "Mais {Amount} para o nível {LevelName}",
   "referral.hw_sales_title": "Vendas mensais",
   "referral.intro_for_you": "Para Você",
   "referral.intro_for_you_1": "Obtenha até {RebateRate} de comissão de venda de carteira de hardware",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Como funciona",
   "referral.invalid_code": "Código de referência inválido",
   "referral.more_questions": "Para mais perguntas, confira <url>https://help.onekey.so/hc/articles/12278189850127<underline>aqui</underline></url>.",
+  "referral.next_distribution": "Próxima distribuição",
+  "referral.next_distribution_day": "Até {Month} 10",
   "referral.no_reward": "Ainda sem recompensa",
   "referral.onboard_bind_code": "Vincular código de indicação",
   "referral.onboard_bind_code_finish": "Terminar",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Recompensa Total",
   "referral.total_reward_pop1": "Recompensas totais confirmadas para todos os convites",
   "referral.total_reward_pop2": "Recompensas pendentes não estão incluídas",
+  "referral.undistributed": "Não distribuído",
   "referral.wallet_code_desc": "Uma vez vinculado, todos os endereços derivados desta carteira serão associados a este código de referência e não poderão ser alterados.",
   "referral.wallet_code_placeholder": "Código de referência",
   "referral.wallet_code_title": "Você tem um código de indicação?",
