+++ packages/shared/src/locale/json/ru.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Ликвидность",
   "global.logout": "Выйти",
   "global.manage": "Управлять",
+  "global.manage_accounts": "Управление аккаунтами",
   "global.mark_all_as_confirmation_desc": "Этого не исправить.",
   "global.mark_all_as_confirmation_title": "Отметить все как прочитанное?",
   "global.market": "Рынок",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Фраза восстановления",
   "global.refresh": "Обновить",
   "global.remove": "Удалить",
-  "global.remove_account_desc": "Вы сможете восстановить этот аккаунт позже в этом кошельке, используя функцию «Массовое добавление аккаунтов».",
+  "global.remove_account_desc": "Вы сможете восстановить этот аккаунт позже в этом кошельке, используя функцию «Управление аккаунтами».",
   "global.remove_account_name": "Удалить {account}",
   "global.rename": "Переименовать",
   "global.reset": "Сброс",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Не можете найти токен? Попробуйте поискать",
   "referral.address_update_desc": "Код был отправлен на {mail}, пожалуйста, подтвердите его перед обновлением адреса.",
   "referral.address_updated": "Адрес получения обновлён",
+  "referral.cumulative_rewards": "Совокупные награды",
+  "referral.distributed": "Распределённый",
   "referral.earn_reward": "Заработайте награду",
   "referral.earn_reward_desc": "Получайте комиссию от своих друзей и зарабатывайте награды.",
+  "referral.earn_reward_details": "Подробности о вознаграждении",
+  "referral.earn_reward_details_desc": "Все токены фиксируются по рыночной стоимости в USDC 1-го числа каждого месяца, а затем распределяются.",
   "referral.earn_reward_tips": "Награды будут распределены на ваш кошелек к 10-му числу следующего месяца.",
   "referral.faqs": "Часто задаваемые вопросы",
   "referral.faqs_a1": "Пригласите друзей использовать ваш реферальный код для покупок Earn или аппаратного кошелька, чтобы получить награды. Награды будут отправлены на ваш адрес для наград примерно через месяц после расчета.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Пригласите друга по вашему реферальному коду",
   "referral.how_3": "Они покупают аппаратный кошелёк или привязывают кошелёк с помощью вашего кода",
   "referral.how_4": "Вы получите награду",
+  "referral.hw_level_rebate_rate": "{percent} вознаграждение",
   "referral.hw_level_title": "Уровень",
+  "referral.hw_level_up_remain": "Осталось {Amount} до уровня {LevelName}",
   "referral.hw_sales_title": "Ежемесячные продажи",
   "referral.intro_for_you": "Для вас",
   "referral.intro_for_you_1": "Получите до {RebateRate} комиссии с продажи аппаратного кошелька",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Как это работает",
   "referral.invalid_code": "Неверный реферальный код",
   "referral.more_questions": "Для получения дополнительной информации, проверьте <url>https://help.onekey.so/hc/articles/12278189850127<underline>здесь</underline></url>.",
+  "referral.next_distribution": "Следующее распределение",
+  "referral.next_distribution_day": "К {Month} 10",
   "referral.no_reward": "Пока нет награды",
   "referral.onboard_bind_code": "Привязать реферальный код",
   "referral.onboard_bind_code_finish": "Заканчивать",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Общая награда",
   "referral.total_reward_pop1": "Общее количество подтвержденных вознаграждений за все приглашения",
   "referral.total_reward_pop2": "Ожидаемые награды не включены",
+  "referral.undistributed": "Нераспределенный",
   "referral.wallet_code_desc": "После привязки все адреса, полученные из этого кошелька, будут связаны с этим реферальным кодом, и это невозможно будет изменить.",
   "referral.wallet_code_placeholder": "Реферальный код",
   "referral.wallet_code_title": "У вас есть реферальный код?",
