+++ packages/shared/src/locale/json/th_TH.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "ความนุ่มนวล",
   "global.logout": "ออกจากระบบ",
   "global.manage": "จัดการ",
+  "global.manage_accounts": "จัดการบัญชี",
   "global.mark_all_as_confirmation_desc": "นี่ไม่สามารถย้อนกลับได้",
   "global.mark_all_as_confirmation_title": "ทำเครื่องหมายทั้งหมดว่าอ่านแล้วหรือไม่?",
   "global.market": "ตลาด",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "วลีการกู้คืน",
   "global.refresh": "รีเฟรช",
   "global.remove": "ลบ",
-  "global.remove_account_desc": "คุณสามารถกู้คืนบัญชีนี้ในกระเป๋าเงินนี้ได้ในภายหลังโดยใช้ \"เพิ่มบัญชีแบบกลุ่ม\"",
+  "global.remove_account_desc": "คุณสามารถกู้คืนบัญชีนี้ในกระเป๋าเงินนี้ได้ในภายหลังโดยใช้ \"จัดการบัญชี\"",
   "global.remove_account_name": "ลบ {account}ออก",
   "global.rename": "เปลี่ยนชื่อ",
   "global.reset": "รีเซ็ต",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "ไม่พบโทเค็นใช่ไหม ลองค้นหาดู",
   "referral.address_update_desc": "รหัสได้ถูกส่งไปยัง {mail} กรุณายืนยันก่อนที่จะอัปเดตที่อยู่",
   "referral.address_updated": "ที่อยู่สำหรับรับได้รับการอัปเดต",
+  "referral.cumulative_rewards": "รางวัลสะสม",
+  "referral.distributed": "กระจาย",
   "referral.earn_reward": "รับรางวัล",
   "referral.earn_reward_desc": "รับค่าคอมมิชชั่นจากเพื่อนของคุณ รับรางวัล",
+  "referral.earn_reward_details": "รายละเอียดรางวัล",
+  "referral.earn_reward_details_desc": "โทเคนทั้งหมดจะถูกจับภาพมูลค่าตลาดเป็น USDC ในวันที่ 1 ของแต่ละเดือน จากนั้นจึงทำการแจกจ่าย",
   "referral.earn_reward_tips": "รางวัลจะถูกแจกจ่ายไปยังวอลเล็ทของคุณภายในวันที่ 10 ของเดือนหน้า",
   "referral.faqs": "คำถามที่พบบ่อย",
   "referral.faqs_a1": "เชิญเพื่อนมาใช้รหัสแนะนำของคุณสำหรับการซื้อ Earn หรือกระเป๋าเงินฮาร์ดแวร์เพื่อรับรางวัล รางวัลจะถูกส่งไปยังที่อยู่รับรางวัลของคุณประมาณหนึ่งเดือนหลังจากการชำระเงินเสร็จสิ้น",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "เชิญเพื่อนของคุณโดยใช้รหัสแนะนำของคุณ",
   "referral.how_3": "พวกเขาซื้อฮาร์ดแวร์วอลเล็ตหรือเชื่อมโยงวอลเล็ตกับโค้ดของคุณ",
   "referral.how_4": "คุณได้รับรางวัล",
+  "referral.hw_level_rebate_rate": "รางวัล {percent}",
   "referral.hw_level_title": "ระดับ",
+  "referral.hw_level_up_remain": "อีก {Amount} เพื่อไปยังระดับ {LevelName}",
   "referral.hw_sales_title": "ยอดขายรายเดือน",
   "referral.intro_for_you": "สำหรับคุณ",
   "referral.intro_for_you_1": "รับค่าคอมมิชชั่นการขายกระเป๋าเงินฮาร์ดแวร์สูงสุด {RebateRate}",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "วิธีการทำงาน",
   "referral.invalid_code": "รหัสอ้างอิงไม่ถูกต้อง",
   "referral.more_questions": "สำหรับคำถามเพิ่มเติม โปรดตรวจสอบ <url>https://help.onekey.so/hc/articles/12278189850127<underline>ที่นี่</underline></url>",
+  "referral.next_distribution": "การแจกจ่ายครั้งถัดไป",
+  "referral.next_distribution_day": "ภายในวันที่ 10 {Month}",
   "referral.no_reward": "ยังไม่มีรางวัล",
   "referral.onboard_bind_code": "ผูกโค้ดแนะนำ",
   "referral.onboard_bind_code_finish": "เสร็จ",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "รางวัลรวม",
   "referral.total_reward_pop1": "รางวัลที่ยืนยันทั้งหมดสำหรับคำเชิญทั้งหมด",
   "referral.total_reward_pop2": "รางวัลที่รอดำเนินการจะไม่ถูกรวม",
+  "referral.undistributed": "ไม่กระจาย",
   "referral.wallet_code_desc": "เมื่อผูกแล้ว ที่อยู่ทั้งหมดที่สร้างจากกระเป๋าเงินนี้จะเชื่อมโยงกับรหัสการแนะนำนี้และไม่สามารถเปลี่ยนแปลงได้",
   "referral.wallet_code_placeholder": "รหัสอ้างอิง",
   "referral.wallet_code_title": "คุณมีรหัสแนะนำหรือไม่?",
