+++ packages/shared/src/locale/json/uk_UA.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Ліквідність",
   "global.logout": "Вийти",
   "global.manage": "Керувати",
+  "global.manage_accounts": "Керування обліковими записами",
   "global.mark_all_as_confirmation_desc": "Цю дію не можна скасувати.",
   "global.mark_all_as_confirmation_title": "Позначити все як прочитане?",
   "global.market": "Ринок",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Фраза відновлення",
   "global.refresh": "Оновити",
   "global.remove": "Видалити",
-  "global.remove_account_desc": "Ви можете відновити цей обліковий запис пізніше в цьому гаманці, використовуючи \\\"Масове додавання облікових записів\\\".",
+  "global.remove_account_desc": "Ви можете відновити цей обліковий запис пізніше в цьому гаманці за допомогою \\\"Керування обліковими записами\\\".",
   "global.remove_account_name": "Видалити {account}",
   "global.rename": "Перейменувати",
   "global.reset": "Скинути",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Не можете знайти маркер? Спробуйте пошукати",
   "referral.address_update_desc": "Код було надіслано на {mail}, будь ласка, підтвердіть перед оновленням адреси.",
   "referral.address_updated": "Адресу отримання оновлено",
+  "referral.cumulative_rewards": "Накопичені винагороди",
+  "referral.distributed": "Розподілений",
   "referral.earn_reward": "Заробляйте винагороду",
   "referral.earn_reward_desc": "Отримуйте комісію від винагород, які заробляють ваші друзі.",
+  "referral.earn_reward_details": "Деталі винагороди",
+  "referral.earn_reward_details_desc": "Усі токени конвертуються в USDC за ринковою вартістю 1-го числа кожного місяця, а потім розподіляються.",
   "referral.earn_reward_tips": "Винагороди будуть розподілені на ваш гаманець до 10-го числа наступного місяця.",
   "referral.faqs": "Часті запитання",
   "referral.faqs_a1": "Запросіть друзів скористатися вашим реферальним кодом для заробітку або покупок апаратного гаманця, щоб отримати винагороди. Винагороди будуть надіслані на вашу адресу для винагород приблизно через місяць після розрахунку.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Запросіть свого друга за допомогою вашого реферального коду",
   "referral.how_3": "Вони купують апаратний гаманець або прив'язують гаманець з вашим кодом",
   "referral.how_4": "Ви отримуєте винагороду",
+  "referral.hw_level_rebate_rate": "{percent} винагорода",
   "referral.hw_level_title": "Рівень",
+  "referral.hw_level_up_remain": "Ще {Amount} до рівня {LevelName}",
   "referral.hw_sales_title": "Щомісячні продажі",
   "referral.intro_for_you": "Для вас",
   "referral.intro_for_you_1": "Отримайте до {RebateRate} комісійних від продажу апаратних гаманців",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Як це працює",
   "referral.invalid_code": "Недійсний реферальний код",
   "referral.more_questions": "Для отримання додаткової інформації перевірте <url>https://help.onekey.so/hc/articles/12278189850127<underline>тут</underline></url>.",
+  "referral.next_distribution": "Наступна дистрибуція",
+  "referral.next_distribution_day": "До 10 {Month}",
   "referral.no_reward": "Ще немає винагороди",
   "referral.onboard_bind_code": "Прив'язати реферальний код",
   "referral.onboard_bind_code_finish": "Фініш",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Загальна винагорода",
   "referral.total_reward_pop1": "Загальна кількість підтверджених винагород за всі запрошення",
   "referral.total_reward_pop2": "Незараховані винагороди не включені",
+  "referral.undistributed": "Нерозподілений",
   "referral.wallet_code_desc": "Після прив'язки всі адреси, отримані з цього гаманця, будуть пов'язані з цим реферальним кодом і не можуть бути змінені.",
   "referral.wallet_code_placeholder": "Реферальний код",
   "referral.wallet_code_title": "У вас є реферальний код?",
