+++ packages/shared/src/locale/json/vi.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "Tính thanh khoản",
   "global.logout": "Đăng xuất",
   "global.manage": "Quản lý",
+  "global.manage_accounts": "Quản lý tài khoản",
   "global.mark_all_as_confirmation_desc": "Không thể hoàn tác được.",
   "global.mark_all_as_confirmation_title": "Đánh dấu tất cả là đã đọc?",
   "global.market": "Thị trường",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "Cụm từ khôi phục",
   "global.refresh": "Làm mới",
   "global.remove": "Xóa",
-  "global.remove_account_desc": "Bạn có thể khôi phục tài khoản này sau trong ví này bằng cách sử dụng \"Thêm nhiều tài khoản\".",
+  "global.remove_account_desc": "Bạn có thể khôi phục tài khoản này sau trong ví này bằng cách sử dụng \"Quản lý tài khoản\".",
   "global.remove_account_name": "Xóa {account}",
   "global.rename": "Đổi tên",
   "global.reset": "Cài lại",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "Không tìm thấy token? Hãy thử tìm kiếm",
   "referral.address_update_desc": "Một mã đã được gửi đến {mail}, vui lòng xác minh trước khi cập nhật địa chỉ.",
   "referral.address_updated": "Địa chỉ nhận đã được cập nhật",
+  "referral.cumulative_rewards": "Phần thưởng tích lũy",
+  "referral.distributed": "Phân tán",
   "referral.earn_reward": "Nhận phần thưởng",
   "referral.earn_reward_desc": "Nhận hoa hồng từ phần thưởng kiếm được của bạn bè.",
+  "referral.earn_reward_details": "Chi tiết phần thưởng",
+  "referral.earn_reward_details_desc": "Tất cả token được chốt giá theo USDC dựa trên giá thị trường vào ngày 1 mỗi tháng, sau đó sẽ được phân phối.",
   "referral.earn_reward_tips": "Phần thưởng sẽ được phân phối vào ví của bạn trước ngày 10 của tháng sau.",
   "referral.faqs": "Câu hỏi thường gặp",
   "referral.faqs_a1": "Mời bạn bè sử dụng mã giới thiệu của bạn để mua Earn hoặc ví phần cứng để nhận phần thưởng. Phần thưởng sẽ được gửi đến địa chỉ phần thưởng của bạn khoảng một tháng sau khi thanh toán.",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "Mời bạn của bạn bằng mã giới thiệu của bạn",
   "referral.how_3": "Họ mua ví phần cứng hoặc liên kết ví bằng mã của bạn",
   "referral.how_4": "Bạn nhận được phần thưởng",
+  "referral.hw_level_rebate_rate": "Phần thưởng {percent}",
   "referral.hw_level_title": "Cấp độ",
+  "referral.hw_level_up_remain": "{Amount} nữa để đạt cấp độ {LevelName}",
   "referral.hw_sales_title": "Doanh số hàng tháng",
   "referral.intro_for_you": "Dành cho bạn",
   "referral.intro_for_you_1": "Nhận hoa hồng bán ví phần cứng lên đến {RebateRate}",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "Cách nó hoạt động",
   "referral.invalid_code": "Mã giới thiệu không hợp lệ",
   "referral.more_questions": "Để biết thêm câu hỏi, hãy kiểm tra <url>https://help.onekey.so/hc/articles/12278189850127<underline>tại đây</underline></url>.",
+  "referral.next_distribution": "Phân phối tiếp theo",
+  "referral.next_distribution_day": "Trước ngày 10 tháng {Month}",
   "referral.no_reward": "Chưa có phần thưởng",
   "referral.onboard_bind_code": "Liên kết mã giới thiệu",
   "referral.onboard_bind_code_finish": "Hoàn thành",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "Tổng Phần Thưởng",
   "referral.total_reward_pop1": "Tổng phần thưởng đã xác nhận cho tất cả lời mời",
   "referral.total_reward_pop2": "Phần thưởng đang chờ xử lý không được bao gồm",
+  "referral.undistributed": "Chưa phân phối",
   "referral.wallet_code_desc": "Sau khi liên kết, tất cả các địa chỉ được tạo từ ví này sẽ được liên kết với mã giới thiệu này và không thể thay đổi.",
   "referral.wallet_code_placeholder": "Mã giới thiệu",
   "referral.wallet_code_title": "Bạn có mã giới thiệu không?",
