+++ packages/shared/src/locale/json/zh_CN.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "流动性",
   "global.logout": "登出",
   "global.manage": "管理",
+  "global.manage_accounts": "管理账户",
   "global.mark_all_as_confirmation_desc": "这无法撤消。",
   "global.mark_all_as_confirmation_title": "全部标记为已读？",
   "global.market": "市场",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "助记词",
   "global.refresh": "刷新",
   "global.remove": "移除",
-  "global.remove_account_desc": "您可以稍后在此钱包中使用「批量添加账户」来恢复此账户。",
+  "global.remove_account_desc": "您可以稍后在此钱包中使用「管理账户」来恢复此账户。",
   "global.remove_account_name": "移除 {account}",
   "global.rename": "重命名",
   "global.reset": "重置",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "找不到代币？试试搜索",
   "referral.address_update_desc": "验证码已发送至 {mail}，请完成验证以更新地址。",
   "referral.address_updated": "收款地址已更新",
+  "referral.cumulative_rewards": "累计奖励",
+  "referral.distributed": "已分配",
   "referral.earn_reward": "理財奖励",
   "referral.earn_reward_desc": "从你朋友的理财奖励中获得佣金。",
+  "referral.earn_reward_details": "奖励详情",
+  "referral.earn_reward_details_desc": "所有代币都会在每月1日按市场价值快照为USDC，然后进行分配。",
   "referral.earn_reward_tips": "奖励将在下个月 10 号之前分发到您的钱包。",
   "referral.faqs": "常见问题",
   "referral.faqs_a1": "邀请好友使用您的邀请码进行理财或购买硬件钱包，即可获得奖励。奖励将在结算后约一个月发送至您的奖励地址。",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "使用邀请码来邀请朋友",
   "referral.how_3": "他们购买硬件钱包或绑定钱包",
   "referral.how_4": "你获得奖励",
+  "referral.hw_level_rebate_rate": "{percent} 奖励",
   "referral.hw_level_title": "等级",
+  "referral.hw_level_up_remain": "还需{Amount}销售额达到{LevelName}级别",
   "referral.hw_sales_title": "本月销售",
   "referral.intro_for_you": "对你",
   "referral.intro_for_you_1": "获得高达 {RebateRate} 的硬件钱包销售佣金",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "如何运作",
   "referral.invalid_code": "邀请码不存在",
   "referral.more_questions": "如有更多问题，请查看<url>https://help.onekey.so/hc/articles/12278189850127<underline>这里</underline></url>。",
+  "referral.next_distribution": "下一次发放",
+  "referral.next_distribution_day": "{Month} 10 日之前",
   "referral.no_reward": "暂无奖励",
   "referral.onboard_bind_code": "绑定邀请码",
   "referral.onboard_bind_code_finish": "完成",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "总奖励",
   "referral.total_reward_pop1": "所有邀请的历史总奖励",
   "referral.total_reward_pop2": "待入账的奖励不包括在内",
+  "referral.undistributed": "未分配",
   "referral.wallet_code_desc": "一旦绑定，从此钱包派生的所有地址将与此邀请码关联，且无法更改。",
   "referral.wallet_code_placeholder": "邀请码",
   "referral.wallet_code_title": "是否填写邀请码？",
