+++ packages/shared/src/locale/json/zh_TW.json
@@ -1263,6 +1263,7 @@
   "global.liquidity": "流動性",
   "global.logout": "登出",
   "global.manage": "管理",
+  "global.manage_accounts": "管理帳戶",
   "global.mark_all_as_confirmation_desc": "這無法撤消。",
   "global.mark_all_as_confirmation_title": "全部標記為已讀？",
   "global.market": "市場",
@@ -1365,7 +1366,7 @@
   "global.recovery_phrase": "助記詞",
   "global.refresh": "刷新",
   "global.remove": "移除",
-  "global.remove_account_desc": "您可以稍後在此錢包中使用「批量添加帳戶」來恢復此帳戶。",
+  "global.remove_account_desc": "您可以稍後在此錢包中使用「管理帳戶」來還原此帳戶。",
   "global.remove_account_name": "移除 {account}",
   "global.rename": "重新命名",
   "global.reset": "重置",
@@ -2110,8 +2111,12 @@
   "receive_token_list_footer_text": "找不到代幣？試試搜尋",
   "referral.address_update_desc": "已將驗證碼發送至 {mail}，請完成驗證以更新地址。",
   "referral.address_updated": "收件地址已更新",
+  "referral.cumulative_rewards": "累積獎勵",
+  "referral.distributed": "已分配",
   "referral.earn_reward": "理財獎勵",
   "referral.earn_reward_desc": "從你朋友的理財獎勵中獲得佣金。",
+  "referral.earn_reward_details": "獎勵詳情",
+  "referral.earn_reward_details_desc": "所有代幣都會在每月1號以市場價值轉換為USDC進行快照，然後進行分發。",
   "referral.earn_reward_tips": "獎勵將在下個月 10 號之前發放到您的錢包。",
   "referral.faqs": "常見問題",
   "referral.faqs_a1": "邀請好友使用您的邀請碼進行理財或購買硬體錢包，即可獲得獎勵。獎勵將在結算後約一個月發送至您的獎勵地址。",
@@ -2125,7 +2130,9 @@
   "referral.how_2": "使用邀請碼邀請朋友",
   "referral.how_3": "他們購買硬體錢包或綁定錢包",
   "referral.how_4": "你獲得獎勵",
+  "referral.hw_level_rebate_rate": "{percent} 獎勵",
   "referral.hw_level_title": "等級",
+  "referral.hw_level_up_remain": "再{Amount}銷售額即可達到{LevelName}等級",
   "referral.hw_sales_title": "本月銷售",
   "referral.intro_for_you": "對你",
   "referral.intro_for_you_1": "獲得高達 {RebateRate} 硬體錢包銷售佣金",
@@ -2138,6 +2145,8 @@
   "referral.intro_title_2": "如何運作",
   "referral.invalid_code": "邀請碼無效",
   "referral.more_questions": "如有更多問題，請查看<url>https://help.onekey.so/hc/articles/12278189850127<underline>這裡</underline></url>。",
+  "referral.next_distribution": "下一次發放",
+  "referral.next_distribution_day": "{Month} 10 日之前",
   "referral.no_reward": "暫無獎勵",
   "referral.onboard_bind_code": "綁定邀請碼",
   "referral.onboard_bind_code_finish": "完成",
@@ -2184,6 +2193,7 @@
   "referral.total_reward": "總獎勵",
   "referral.total_reward_pop1": "所有邀請的歷史總獎勵",
   "referral.total_reward_pop2": "待入帳的獎勵不包括在內",
+  "referral.undistributed": "未分配",
   "referral.wallet_code_desc": "一旦綁定，從此錢包衍生的所有地址都將與此邀請碼關聯，且無法更改。",
   "referral.wallet_code_placeholder": "邀請碼",
   "referral.wallet_code_title": "是否要填寫邀請碼？",
